+++ tools/solidity-to-uml.mjs
@@ -25,9 +25,15 @@ await generateClassDiagramOfContracts(
         'IexecEscrowNativeDelegate',
         'IexecEscrowTokenDelegateKYC',
         'IexecEscrowTokenDelegate',
-        'IexecEscrowTokenSwapDelegate'
+        'IexecEscrowTokenSwapDelegate',
+        'IexecEscrow',
     ],
     'IexecEscrows',
+)
+
+await generateClassDiagramOfContracts(
+    ['IexecPocoBoostDelegate'],
+    'IexecPocoBoostDelegate',
 )
 
 /**